requirements.txt
python-keystoneclient>=1.1.0,<1.2
requests>=2.5.3,<2.6
